Address code review: fix potential overflow and add path sanitization

+++ app/src/main/java/com/hdclark/gpxlogger/MainActivity.kt
@@ -297,7 +297,7 @@ class MainActivity : AppCompatActivity() {
     
     private fun updateElapsedTime() {
         if (lastLocationUpdateTime > 0) {
-            val elapsedSeconds = ((System.currentTimeMillis() - lastLocationUpdateTime) / MILLIS_PER_SECOND).toInt()
+            val elapsedSeconds = (System.currentTimeMillis() - lastLocationUpdateTime) / MILLIS_PER_SECOND
             lastUpdateText.text = getString(R.string.seconds_ago, elapsedSeconds)
         }
     }
